Split resolver files and fixed manifesto accordingly
